QuantasticaAPI/NewsScrapper/requirements.txt
Flask==2.3.3
newspaper3k==0.2.8
newsapi-python==0.2.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml_html_clean
